+++ internal/topic/topicwriterinternal/writer_transaction.go
@@ -2,7 +2,6 @@ package topicwriterinternal
 
 import (
 	"context"
-	"fmt"
 
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/topic/gtrace"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/tx"
@@ -60,10 +59,6 @@ func (w *WriterWithTransaction) WaitInitInfo(ctx context.Context) (InitialInfo, 
 }
 
 func (w *WriterWithTransaction) Write(ctx context.Context, messages []PublicMessage) error {
-	if err := w.tx.UnLazy(ctx); err != nil {
-		return fmt.Errorf("ydb: failed to materialize transaction: %w", err)
-	}
-
 	for i := range messages {
 		messages[i].Tx = w.tx
 	}